+++ src/lib/resolvers/scss.ts
@@ -1,8 +1,6 @@
-import { resolveHttpUrl, resolveLocalUrl, resolveString } from 'hercule';
-import * as path from 'path';
 import * as fs from 'fs';
 import * as cache from 'js-cache';
-import * as toString from 'stream-to-string';
+import * as path from 'path';
 import { promisify } from 'util';
 
 type SassImportRequest = {
@@ -25,42 +23,11 @@ const read = promisify(fs.readFile);
 const exists = promisify(fs.exists);
 
 /**
- * Basically, we're caching file transcludes so we don't make
- * a bajillion fs requests with every setState.
- *
- * @todo - add to some sort of fs watchlist?
- */
-
-export function markdown(base: string) {
-  return [
-    resolveHttpUrl,
-    resolveLocalUrl,
-    resolveString
-  ].map(resolve => {
-    return (url: string) => {
-      const has = cache.get(url);
-      if (has) {
-        return has;
-      } else {
-        const res = resolve(path.resolve(base, url), '/');
-        if (res) {
-          toString(res.content).then((res: string) =>
-            cache.set(url, { url, content: res })
-          );
-          return res;
-        }
-      }
-      return null;
-    };
-  });
-}
-
-/**
  * Resolve SASS imports.
  *
  * @todo ditto with the watchlist...
  */
-export function sass(base: () => string) {
+export default function sass(base: () => string) {
   return async (
     request: SassImportRequest,
     done: SassImportRequestCallback